changes made per tiago18c review
